requirements.txt
scikit-learn>=1.3.0
PyYAML>=6.0
pyserial>=3.5
joblib>=1.3.0
pandas>=2.0.0
